Add method to check if array contains only scalable values
